已经调通整个流程

+++ src/main/java/com/bigdata/dywz/validate/ValidateReducer.java
@@ -38,7 +38,7 @@ public class ValidateReducer extends Reducer<NullWritable, Text, DoubleWritable,
         }
 
         //计算正确率
-        double accuracy = sum/count;
+        double accuracy = (double) sum/count;
         context.write(new DoubleWritable(accuracy), NullWritable.get());
     }
 }
